update ExchangeRateController, Dto, Form and added log

+++ src/main/java/toyproject/exchangerate/dto/ExchangeRateDto.java
@@ -4,12 +4,14 @@ import lombok.Getter;
 import lombok.Setter;
 import toyproject.exchangerate.data.CountryCode;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ExchangeRateDto {
     private CountryCode countryCode;
     private double exchangeRate;
-    private double receptionAmount;
+    private BigDecimal receptionAmount;
     private int remittanceAmount;
     boolean status = false;
 }
